requirements.txt
streamlit>=1.41.0
anthropic>=0.42.0
transformers>=4.47.0
torch>=2.1.0
requests>=2.32.0
python-dotenv>=1.0.0
PyMuPDF>=1.25.0
reportlab>=4.2.0
pandas>=2.2.0
plotly>=5.24.0
